Rename subfolder migrations
