+++ openslides_backend/http/views/action_view.py
@@ -1,8 +1,8 @@
 from typing import Optional, Tuple
 
 from ...action.action_handler import ActionHandler
-from ...migrations import assert_migration_index
-from ...migrations.migration_handler import MigrationHandler
+from ...migration_handler import assert_migration_index
+from ...migration_handler.migration_handler import MigrationHandler
 from ...shared.interfaces.wsgi import ResponseBody
 from ..request import Request
 from .base_view import BaseView, route
